Disallow user to select a blank vibe in the carousel
