+++ docs/javadocs/com/example/vybe/VibeFactoryTest.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.vybe.VibeFactoryTest</li>
+<li>com.example.vybe.VibeTest</li>
 </ul>
 </li>
 </ul>
